feat(@uform/react): rewrite shared/core
